Update split_dpm_bpm_fq.py to be more robust
Prior to the previous commit, split_dpm_bpm_fq.py would have dump reads with tags not specified in the format file into <sample>_other.fastq.gz.

The previous commit added an implicit assumption assumed that all tags encountered would be in format.txt. Consequently, the updated pipeline threw an error when it encountered a tag that BarcodeID had identified but was not specified in the format file.

This update removes the assumption that all tags encountered would be in the format file, thereby reverting to the old behavior (dump reads to <sample>_other.fastq.gz ) if it encounters a tag not specified in format file, but retain the new behavior of allowing tags to be re-used across rounds.