Fix bug: split_fastq.sh could not always generate requested number of chunks
Problem: When requesting a large number of chunks for a small number of reads, split_fastq.sh could not always create the exact number of chunks requested (the number created would be less than requested). Since the pipeline expects the exact number of chunks to be generated, failure to generate all the chunks (files) would cause snakemake to report an error indicating that the split_fastq rule failed to generate the expected output files.
- Specifically, split_fastq.sh computes the number of reads to place into each chunk. This number is then passed to GNU split in the form split --lines=<# lines>, where the # of lines is just 4 * # of reads.
- Example where this usage fails to generate the number of requested chunks: 3 chunks requested for 4 reads (16 lines)
  - split --lines=4 --> 4 chunks of 1 read = 4 total chunks
  - split --lines=8 --> 2 chunks of 2 reads = 2 total chunks
  - split --lines=12 --> 1 chunk of 3 reads, 1 chunk of 1 read = 2 total chunks

Solution: When GNU split cannot generate the number of chunks requested, use GNU csplit instead, which is also part of GNU coreutils, but which runs slower than split.
- One potential downside to using csplit instead of split is that it does not support the --filter=command argument, which precludes piping the splits directly into pigz or another downstream command.