Bug fix: error when sorting required in rename_and_filter_chr.py and output to standard out
Background: When sorting is required, rename_and_filter_chr.py:filter_reads() creates a
subprocess for samtools sort and pipes reads (lines) to the subprocess.
When no specific output file path is supplied, the result should be written to
standard output.

Problem: Previously, the stdout argument of the Popen constructor for this subprocess
was set to sys.stdout.buffer if no specific output file path was supplied. This would
raise an exception.

Fix: Instead, the stdout argument should just be set to None. As
mentioned in the Python subprocess library documentation:
> With the default settings of None, no redirection will occur.