Bug fix: repeat_mask rule would fail if masking disabled
Problem: If the `mask` key was set to null or "" in config/config.yaml,
then the repeat_mask rule would create a broken symbolic link between
the input and output files.

Solution: The source path for the symbolic link needs to be provided either
as a relative path *relative to the output path* or as an absolute path.
Here, we use an absolute path.

Credits to Drew Honson (@honsonbiosci) for pointing out this bug.